+++ nemo/collections/common/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from nemo.collections.common import tokenizers, parts
+from nemo.collections.common import tokenizers
